refactor(goal): extract pure goal subsystems
