+++ extensions/pi-goal/src/persistence.ts
@@ -1,0 +1,104 @@
+import { existsSync, mkdirSync, readFileSync, writeFileSync } from "node:fs";
+import { dirname, join } from "node:path";
+import process from "node:process";
+import {
+	isNonNegativeFiniteNumber,
+	nonNegativeFiniteNumber,
+	normalizeTokenBudget,
+} from "./accounting.js";
+import type { GoalStatus } from "./prompts.js";
+
+const GOAL_STATE_ENTRY_TYPE = "goal-state";
+const STATE_FILE = join(
+	process.env.PI_CODING_AGENT_DIR ?? join(process.env.HOME ?? ".", ".pi", "agent"),
+	"pi-goal-state.json",
+);
+
+export interface ActiveGoal {
+	id: string;
+	text: string;
+	status: GoalStatus;
+	startedAt: number;
+	updatedAt: number;
+	iteration: number;
+	tokenBudget?: number;
+	tokensUsed: number;
+	timeUsedSeconds: number;
+	baselineTokens: number;
+	activeStartedAt?: number;
+}
+
+export interface GoalStateEntryData {
+	goal: ActiveGoal | null;
+}
+
+interface SessionContext {
+	sessionManager?: {
+		getBranch?: () => Array<{ type?: string; customType?: string; data?: unknown }>;
+		getEntries?: () => Array<{ type?: string; customType?: string; data?: unknown }>;
+	};
+}
+
+export function loadGoalFromSession(ctx: SessionContext): ActiveGoal | undefined {
+	const entries = ctx.sessionManager?.getBranch?.() ?? ctx.sessionManager?.getEntries?.() ?? [];
+	const entry = entries
+		.filter((entry) => entry.type === "custom" && entry.customType === GOAL_STATE_ENTRY_TYPE)
+		.pop();
+	const data = entry?.data as GoalStateEntryData | undefined;
+	if (!isGoal(data?.goal) || data.goal.status === "complete") return undefined;
+	return normalizeLoadedGoal(data.goal);
+}
+
+export function normalizeLoadedGoal(goal: ActiveGoal): ActiveGoal {
+	const now = Date.now();
+	return {
+		...goal,
+		startedAt: isNonNegativeFiniteNumber(goal.startedAt) ? goal.startedAt : now,
+		updatedAt: isNonNegativeFiniteNumber(goal.updatedAt) ? goal.updatedAt : now,
+		iteration: Math.max(0, Math.floor(nonNegativeFiniteNumber(goal.iteration))),
+		tokenBudget: normalizeTokenBudget(goal.tokenBudget),
+		tokensUsed: nonNegativeFiniteNumber(goal.tokensUsed),
+		timeUsedSeconds: nonNegativeFiniteNumber(goal.timeUsedSeconds),
+		baselineTokens: nonNegativeFiniteNumber(goal.baselineTokens),
+		activeStartedAt: goal.status === "active" ? now : undefined,
+	};
+}
+
+export function clearLegacyPersistedGoal(cwd: string) {
+	if (!existsSync(STATE_FILE)) return;
+	const goals = readState();
+	delete goals[cwd];
+	mkdirSync(dirname(STATE_FILE), { recursive: true });
+	writeFileSync(STATE_FILE, `${JSON.stringify(goals, null, 2)}\n`);
+}
+
+function readState(): Record<string, unknown> {
+	if (!existsSync(STATE_FILE)) return {};
+	try {
+		const parsed = JSON.parse(readFileSync(STATE_FILE, "utf8")) as unknown;
+		return parsed && typeof parsed === "object" && !Array.isArray(parsed)
+			? (parsed as Record<string, unknown>)
+			: {};
+	} catch {
+		return {};
+	}
+}
+
+function isGoal(value: unknown): value is ActiveGoal {
+	if (!value || typeof value !== "object") return false;
+	const goal = value as Partial<ActiveGoal>;
+	return (
+		typeof goal.id === "string" &&
+		typeof goal.text === "string" &&
+		["active", "paused", "blocked", "usage_limited", "budget_limited", "complete"].includes(
+			String(goal.status),
+		) &&
+		typeof goal.startedAt === "number" &&
+		typeof goal.updatedAt === "number" &&
+		typeof goal.iteration === "number" &&
+		typeof goal.tokensUsed === "number" &&
+		typeof goal.timeUsedSeconds === "number" &&
+		typeof goal.baselineTokens === "number" &&
+		(goal.activeStartedAt === undefined || typeof goal.activeStartedAt === "number")
+	);
+}
